# services/jenny-api/src/server-utfa.ts
import * as dotenv from 'dotenv';
import * as path from 'path';
import { fileURLToPath } from 'url';

const __filename = fileURLToPath(import.meta.url);
const __dirname = path.dirname(__filename);
dotenv.config({ path: path.resolve(__dirname, '../../../.env') });

// Override with correct index
process.env.PINECONE_INDEX = 'jenny-v3-3072-093025';
process.env.SERVICE_NAME = 'jenny-api-utfa';

import express from 'express';
import { agentChat } from './orchestrator/agentChat-utfa.js';
import { fetchVitals, getStudentFactValidation } from './services/facts-canonical.js';
import { fetchLifecycle } from './services/lifecycle.js';
import { resolveTemporalFact } from './services/temporalFacts.js';
import { getInitialAwardTargets, getAwardTargetsByPhase, getAwardTargetsAsOf } from './resolvers/awards.js';
import { EnumerationResolver, FactsResolver } from './resolvers/kb-items.js';
import { pool } from './db/pool.js';
import { enumsRouter } from './routes/enums.js';

const app = express();

// Initialize resolvers
const enumResolver = new EnumerationResolver(pool);
const factsResolver = new FactsResolver(pool);

// Enable CORS
app.use((req, res, next) => {
  res.header('Access-Control-Allow-Origin', '*');
  res.header('Access-Control-Allow-Methods', 'GET, POST, PUT, DELETE, OPTIONS');
  res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept, Authorization');
  if (req.method === 'OPTIONS') {
    res.sendStatus(200);
  } else {
    next();
  }
});

app.use(express.json());

// Mount universal enumerations router
app.use('/enum', enumsRouter(pool));

// Health routes
app.get('/health', (_req, res) => {
  res.json({ ok: true });
});

app.get('/health/details', async (_req, res) => {
  res.json({
    ok: true,
    index_name: process.env.PINECONE_INDEX,
    temporal_facts: 'UTFA',
    db_ping_ms: -1,
    uptime_s: Math.floor(process.uptime())
  });
});

// UTFA test endpoints
app.get('/utfa/test/:student_id', async (req, res) => {
  try {
    const { student_id } = req.params;
    const { kind = 'sat_total_score' } = req.query;
    
    const tests = await Promise.all([
      resolveTemporalFact(pool, { student_id, kind: String(kind), operator: 'first' }),
      resolveTemporalFact(pool, { student_id, kind: String(kind), operator: 'nth', nth: 2 }),
      resolveTemporalFact(pool, { student_id, kind: String(kind), operator: 'latest' }),
      resolveTemporalFact(pool, { student_id, kind: String(kind), operator: 'series' })
    ]);
    
    res.json({
      student_id,
      kind,
      first: tests[0].facts[0] || null,
      second: tests[1].facts[0] || null,
      latest: tests[2].facts[0] || null,
      series: tests[3].facts
    });
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
});

// Award targets endpoint
app.get('/students/:id/awards/targets', async (req, res) => {
  try {
    const { id: studentId } = req.params;
    const { phase, as_of } = req.query;
    
    let result;
    
    if (as_of) {
      const asOfDate = new Date(String(as_of));
      result = await getAwardTargetsAsOf(pool, studentId, asOfDate);
    } else if (phase) {
      result = await getAwardTargetsByPhase(pool, studentId, phase as any);
    } else {
      // Default to initial
      result = await getInitialAwardTargets(pool, studentId);
    }
    
    if (result.count === 0) {
      res.status(412).json({ 
        error: 'no_evidence', 
        message: 'No award targets found',
        student_id: studentId,
        phase: phase || 'initial'
      });
    } else {
      res.json(result.targets);
    }
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
});

// Student routes
app.get('/students/:id/vitals', async (req, res) => {
  try {
    const result = await fetchVitals(req.params.id);
    res.json(result);
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
});

app.get('/students/:id/fact-validation', async (req, res) => {
  try {
    const result = await getStudentFactValidation(req.params.id);
    res.json({ student_id: req.params.id, validation: result });
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
});

app.get('/students/:id/lifecycle', async (req, res) => {
  try {
    const result = await fetchLifecycle(req.params.id);
    res.json(result);
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
});

// Agent chat route with UTFA
app.post('/agent/chat', async (req, res) => {
  try {
    console.log('[UTFA] Chat request:', { 
      message: req.body.message, 
      student_id: req.body.student_id,
      temporal: 'UTFA' 
    });
    
    const result = await agentChat(req.body, res);
    
    if (!req.body?.stream) {
      res.json(result);
    }
  } catch (error: any) {
    console.error('Chat error:', error);
    res.status(500).json({ error: error.message });
  }
});

// KB Items endpoints (Universal Ledger)
app.get('/students/:id/awards/initial', async (req, res) => {
  try {
    const result = await enumResolver.awardsInitial(req.params.id);
    res.json(result);
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
});

app.get('/students/:id/awards/won', async (req, res) => {
  try {
    const result = await enumResolver.awardsWon(req.params.id);
    res.json(result);
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
});

app.get('/students/:id/awards/timeline', async (req, res) => {
  try {
    const result = await enumResolver.awardsTimeline(req.params.id);
    res.json(result);
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
});

// SAT progression endpoints
app.get('/students/:id/testing/sat/first', async (req, res) => {
  try {
    const result = await factsResolver.satFirst(req.params.id);
    res.json(result);
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
});

app.get('/students/:id/testing/sat/latest', async (req, res) => {
  try {
    const result = await factsResolver.satLatest(req.params.id);
    res.json(result);
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
});

app.get('/students/:id/testing/sat/n/:n', async (req, res) => {
  try {
    const result = await factsResolver.satNth(req.params.id, parseInt(req.params.n));
    res.json(result);
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
});

app.get('/students/:id/testing/sat/all', async (req, res) => {
  try {
    const result = await factsResolver.satAll(req.params.id);
    res.json(result);
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
});

// Evidence resolution route
app.get('/evidence', async (req, res) => {
  try {
    const ids = req.query.ids ? String(req.query.ids).split(',') : [];
    console.log('Evidence request:', { ids });

    // For now, return mock evidence data
    const evidence = ids.map(id => ({
      source_id: id,
      title: `Evidence ${id}`,
      content: `This is evidence content for ${id}`,
      drive_link: null,
      type: 'source'
    }));

    res.json(evidence);
  } catch (error: any) {
    console.error('Evidence error:', error);
    res.status(500).json({ error: error.message });
  }
});

// Error handler
app.use((err: any, _req: any, res: any, _next: any) => {
  console.error('Server error:', err);
  res.status(500).json({ error: 'Internal server error' });
});

const port = process.env.PORT || 8787;
app.listen(port, () => {
  console.log(`
============================================
Jenny v3 API with UTFA (Universal Temporal Facts Architecture)
============================================
Port: ${port}
Index: ${process.env.PINECONE_INDEX}
Temporal Facts: UTFA ENABLED ✓

UTFA provides deterministic temporal resolution for:
- First/initial/earliest facts
- Second, third, nth facts  
- Latest/final/last facts
- As-of date queries
- Complete series/history
- Superscore calculations
- Official vs practice filtering

Works universally for:
- SAT/ACT/AP scores
- Awards (applied/won)
- GPAs over time
- Applications/submissions
- Any temporal fact type

Test endpoints:
  GET /utfa/test/:student_id?kind=sat_total_score

Available routes:
  GET  /health
  GET  /health/details
  GET  /students/:id/vitals
  GET  /students/:id/fact-validation
  GET  /students/:id/lifecycle
  POST /agent/chat
  GET  /evidence

KB Items (Universal Ledger):
  GET  /students/:id/awards/initial      - Initial award targets (Planned/Targeted)
  GET  /students/:id/awards/won          - Awards won (Outcome state)
  GET  /students/:id/awards/timeline     - Full awards timeline

SAT Progression:
  GET  /students/:id/testing/sat/first   - First SAT score
  GET  /students/:id/testing/sat/latest  - Latest SAT score
  GET  /students/:id/testing/sat/n/:n    - Nth SAT score
  GET  /students/:id/testing/sat/all     - All SAT scores

Try these queries:
- "What was my first SAT score?"
- "What was my second SAT score?"  
- "Show me all my SAT scores"
- "What was my SAT score as of March 2024?"
- "What's my SAT superscore?"
============================================
`);
});